requirements.txt
playwright>=1.40
lxml>=4.9
asyncpg>=0.27
python-dotenv>=1.0
aiofiles>=23.2
aiohttp>=3.9
pytest>=7.0